refactor(html):   - split tests into several files   - lift up options to factory level   - small performance optimisations
