+++ src/html.js
@@ -77,8 +77,8 @@ function getChunkType(chunk){
   return 'text';
 }
 
-function setAttribute($target, attributeName, attributeValue, originalValue){
-  if(originalValue !== '' && attributeName !== 'style' && attributeName in $target){
+function setAttribute($target, attributeName, attributeValue){
+  if(attributeValue !== '' && attributeName !== 'style' && attributeName in $target){
     $target[attributeName] = attributeValue;
     return;
   }
@@ -87,9 +87,11 @@ function setAttribute($target, attributeName, attributeValue, originalValue){
 
 const stopNodeValue = `modulor_stop_node_${+(new Date())}`;
 
-export function stopNode($target){
-  $target.nodeStopper = stopNodeValue;
-};
+/**
+ *  stopNode directive
+ *  @deprecated
+ * */
+export function stopNode(){};
 
 const DEFAULT_PREFIX  = `{modulor_html_chunk_${+new Date()}:`;
 const DEFAULT_POSTFIX = '}';
@@ -100,397 +102,422 @@ const DEFAULT_SANITIZE_NODE_PREFIX = `modulor_sanitize_node_${+(new Date())}:`;
 const sanitizeTags = ['table', 'tr', 'td'];
 const sanitizeTagsRegex = new RegExp(`<([ /])?(${sanitizeTags.join('|')})([ ][^]>)?`, 'igm');
 
-export function Template(options){
-  this.PREFIX = options.PREFIX || DEFAULT_PREFIX;
-  this.POSTFIX = options.POSTFIX || DEFAULT_POSTFIX;
-
-  this.parser = options.parser || DEFAULT_PARSER;
-
-  this.splitChunkRegex = new RegExp(this.getTokenRegExp(), 'ig');
-  this.findChunksRegex = new RegExp(this.getTokenRegExp(), 'ig');
-  this.replaceChunkRegex = new RegExp(this.getTokenRegExp(true), 'ig');
-  this.matchChunkRegex = new RegExp(`^${this.getTokenRegExp(true)}$`);
-
-  this.sanitizeNodePrefix = options.SANITIZE_NODE_PREFIX || DEFAULT_SANITIZE_NODE_PREFIX;
-
-  this.updates = [];
-
-  return this;
-};
 
 export const containersMap = new Map();
 
-Template.prototype.parse = function(chunks, ...interpolations){
-  this.prevValues = [];
-  this.values = interpolations;
+export function createHtml(options = {}){
 
-  this.template = this.prepareLiterals(chunks);
+  const PREFIX = options.PREFIX || DEFAULT_PREFIX;
+  const POSTFIX = options.POSTFIX || DEFAULT_POSTFIX;
 
-  this.templateId = hash(this.template);
-  const cached = templatesCache[this.templateId];
+  const parser = options.parser || DEFAULT_PARSER;
 
-  if(!isDefined(cached)){
-    this.container = this.generateContainer(this.sanitize(this.template));
-    templatesCache[this.templateId] = this.container;
-  } else {
-    this.container = cached;
+  const splitChunkRegex = new RegExp(getTokenRegExp(), 'ig');
+  const findChunksRegex = new RegExp(getTokenRegExp(), 'ig');
+  const replaceChunkRegex = new RegExp(getTokenRegExp(true), 'ig');
+  const matchChunkRegex = new RegExp(`^${getTokenRegExp(true)}$`);
+
+  const sanitizeNodePrefix = options.SANITIZE_NODE_PREFIX || DEFAULT_SANITIZE_NODE_PREFIX;
+
+  function getChunkById(id, dataMap = []){
+    return dataMap[id];
+  };
+
+  function replaceTokens(text, dataMap = []){
+    return text.replace(replaceChunkRegex, (token, _, index) => {
+      const chunk = getChunkById(index, dataMap);
+      return isDefined(chunk) ? chunk : '';
+    });
+  };
+
+  function processTextNodeChunks(chunks = []){
+    return {
+      childNodes: chunks.map((chunk) => ({
+        nodeType: NODE_TYPES.TEXT_RESULT_NODE,
+        textContent: chunk
+      }))
+    };
+  };
+
+  function generateContainer(markup){
+    return parser.parseFromString(markup, "text/html").body;
+  };
+
+  function generateTokenName(index){
+    return `${PREFIX}${index}${POSTFIX}`;
+  };
+
+  function prepareLiterals([firstChunk, ...restChunks]){
+    return restChunks.reduce((acc, chunk, index) => {
+      const keyName = generateTokenName(index);
+      return acc.concat(keyName).concat(chunk);
+    }, firstChunk);
+  };
+
+  function getTokenRegExp(groupMatches){
+    const indexRegex = `${groupMatches ? '(' : ''}\\d+${groupMatches ? ')' : ''}`;
+    return `(${regExpEscape(PREFIX)}${indexRegex}${regExpEscape(POSTFIX)})`;
+  };
+  
+  function sanitize(str){
+    return str.replace(sanitizeTagsRegex, `<$1${sanitizeNodePrefix}$2`);
+  };
+
+  function copyTextNodeChunks(chunks){
+    const container = {
+      childNodes: [].concat(chunks).reduce((acc, chunk) => {
+
+        if(!isDefined(chunk)){
+          return acc;
+        }
+
+        const chunkType = getChunkType(chunk);
+
+        const $el = {};
+
+        if(chunkType === 'text'){
+          $el.nodeType = NODE_TYPES.TEXT_RESULT_NODE;
+          $el.textContent = chunk;
+        }
+        if(chunkType === 'element'){
+          $el.nodeType = NODE_TYPES.ELEMENT_RESULT_NODE;
+        }
+        if(chunkType === 'template'){
+          $el.nodeType = NODE_TYPES.TEMPLATE_NODE;
+        }
+        if(chunkType === 'futureResult'){
+          $el.nodeType = NODE_TYPES.PROMISE_NODE;
+        }
+
+        $el[chunkType] = chunk;
+
+        return acc.concat($el);
+      }, [])
+    };
+    return container;
+  };
+
+  function html(...args){
+    return new Template(...args);
   }
-  return this;
-}
 
-Template.prototype.getChunkById = function(id, dataMap = this.values){
-  return dataMap[id];
-}
+  function render(template, target){
+    return template.render(target);
+  }
 
-Template.prototype.processTextNodeChunks = function(chunks = []){
-  return {
-    childNodes: chunks.map((chunk) => ({
-      nodeType: NODE_TYPES.TEXT_RESULT_NODE,
-      textContent: chunk
-    }))
-  };
-};
+  function Template(chunks = [], ...interpolations){
+    this.prevValues = [];
+    this.values = interpolations;
 
-Template.prototype.copyTextNodeChunks = function(chunks){
-  const container = {
-    childNodes: [].concat(chunks).reduce((acc, chunk) => {
+    this.template = prepareLiterals(chunks);
 
-      if(!isDefined(chunk)){
-        return acc;
-      }
-
-      const chunkType = getChunkType(chunk);
-
-      const $el = {};
-
-      if(chunkType === 'text'){
-        $el.nodeType = NODE_TYPES.TEXT_RESULT_NODE;
-        $el.textContent = chunk;
-      }
-      if(chunkType === 'element'){
-        $el.nodeType = NODE_TYPES.ELEMENT_RESULT_NODE;
-      }
-      if(chunkType === 'template'){
-        $el.nodeType = NODE_TYPES.TEMPLATE_NODE;
-      }
-      if(chunkType === 'futureResult'){
-        $el.nodeType = NODE_TYPES.PROMISE_NODE;
-      }
-
-      $el[chunkType] = chunk;
-      $el.templateId = this.templateId;
-
-      return acc.concat($el);
-    }, [])
-  };
-  return container;
-};
-
-Template.prototype.copyAttributes = function(target, source){
-  const sourceAttributes = source.attributes;
-  const targetAttributes = target.attributes;
-
-  const updates = [];
-  const attributesToKeep = {};
-
-  for(let i = 0; i < sourceAttributes.length; i++){
-    let { name, value }  = sourceAttributes[i];
-    const nameIsDynamic = name.match(this.findChunksRegex);
-    const valueIsDynamic = value.match(this.findChunksRegex);
-
-    attributesToKeep[name] = true;
-
-    if(!nameIsDynamic && !valueIsDynamic){
-      setAttribute(target, name, value);
-      continue;
+    if(!this.template){
+      return this;
     }
 
-    if(name === 'class'){
-      target.className = '';
-      value.split(' ').forEach((className) => {
-        if(!className.match(this.findChunksRegex)){
-          target.classList.add(className);
+    this.templateId = hash(this.template);
+    const cached = templatesCache[this.templateId];
+
+    if(!isDefined(cached)){
+      this.container = generateContainer(sanitize(this.template));
+      templatesCache[this.templateId] = this.container;
+    } else {
+      this.container = cached;
+    }
+
+    this.updates = [];
+
+    return this;
+  };
+
+  Template.prototype.render = function(target = document.createDocumentFragment()){
+    const cached = containersMap.get(target);
+    if((cached || {}).templateId === this.templateId){
+      cached.update(this.values);
+      return;
+    }
+    containersMap.set(target, {
+      templateId: this.templateId,
+      update: this.update.bind(this),
+    });
+    return this.loop(this.container, target);
+  };
+
+  Template.prototype.update = function(newData){
+    this.prevValues = this.values;
+    this.values = newData;
+    this.updates.forEach(u => u());
+  };
+
+  Template.prototype.copyAttributes = function(target, source){
+    const sourceAttributes = source.attributes;
+    const targetAttributes = target.attributes;
+
+    const updates = [];
+    const attributesToKeep = {};
+
+    for(let i = 0; i < sourceAttributes.length; i++){
+      const { name, value }  = sourceAttributes[i];
+      const nameIsDynamic = name.match(findChunksRegex);
+      const valueIsDynamic = value.match(findChunksRegex);
+
+      attributesToKeep[name] = true;
+
+      if(!nameIsDynamic && !valueIsDynamic){
+        setAttribute(target, name, value);
+        continue;
+      }
+
+      if(name === 'class'){
+        target.className = '';
+        value.split(' ').forEach((className) => {
+          if(!className.match(findChunksRegex)){
+            target.classList.add(className);
+            return;
+          }
+          updates.push(() => {
+            const newValue = replaceTokens(className, this.values);
+            const oldValue = replaceTokens(className, this.prevValues);
+            if(oldValue !== newValue){
+              oldValue && target.classList.remove(oldValue);
+              newValue && target.classList.add(newValue);
+            }
+          });
+        });
+
+        continue;
+      }
+
+      const matchName = name.match(matchChunkRegex);
+      const matchValue = value.match(matchChunkRegex);
+
+      updates.push(() => {
+        const preparedName = matchName ? getChunkById(matchName[2], this.values) : replaceTokens(name, this.values);
+        const preparedPrevName = matchName ? getChunkById(matchName[2], this.prevValues) : replaceTokens(name, this.prevValues);
+
+        const preparedValue = matchValue ? getChunkById(matchValue[2], this.values) : replaceTokens(value, this.values);
+        const preparedPrevValue = matchValue ? getChunkById(matchValue[2], this.prevValues) : replaceTokens(value, this.prevValues);
+
+        if(preparedName === preparedPrevName && preparedValue === preparedPrevValue){
           return;
         }
-        updates.push(() => {
-          const newValue = this.replaceTokens(className);
-          const oldValue = this.replaceTokens(className, this.prevValues);
-          if(oldValue !== newValue){
-            oldValue && target.classList.remove(oldValue);
-            newValue && target.classList.add(newValue);
-          }
-        });
+
+        if(preparedName !== preparedPrevName){
+          target.removeAttribute(preparedPrevName);
+        }
+
+        if(!preparedName){
+          return;
+        }
+
+        if(getChunkType(preparedName) === 'function'){
+          preparedName(target, preparedValue);
+          return;
+        }
+
+        if(getChunkType(preparedValue) === 'futureResult'){
+          preparedValue.then((result) => setAttribute(target, preparedName, result));
+          return;
+        }
+
+        setAttribute(target, preparedName, preparedValue);
       });
-
-      continue;
     }
 
-    const matchName = name.match(this.matchChunkRegex);
-    const matchValue = value.match(this.matchChunkRegex);
 
-    updates.push(() => {
-      const preparedName = matchName ? this.getChunkById(matchName[2]) : this.replaceTokens(name);
-      const preparedPrevName = matchName ? this.getChunkById(matchName[2], this.prevValues) : this.replaceTokens(name, this.prevValues);
-
-      const preparedValue = matchValue ? this.getChunkById(matchValue[2]) : this.replaceTokens(value);
-      const preparedPrevValue = matchValue ? this.getChunkById(matchValue[2], this.prevValues) : this.replaceTokens(value, this.prevValues);
-
-      if(preparedName === preparedPrevName && preparedValue === preparedPrevValue){
-        return;
+    for(let i = 0; i < targetAttributes.length; i++){
+      let { name } = targetAttributes[i];
+      if(!attributesToKeep[name]){
+        target.removeAttribute(name);
       }
-
-      if(preparedName !== preparedPrevName){
-        target.removeAttribute(preparedPrevName);
-      }
-
-      if(!preparedName){
-        return;
-      }
-
-      if(getChunkType(preparedName) === 'function'){
-        preparedName(target, preparedValue);
-        return;
-      }
-
-      if(getChunkType(preparedValue) === 'futureResult'){
-        preparedValue.then((result) => setAttribute(target, preparedName, result));
-        return;
-      }
-
-      setAttribute(target, preparedName, preparedValue, value);
-    });
-  }
-
-
-  for(let i = 0; i < targetAttributes.length; i++){
-    let { name } = targetAttributes[i];
-    if(!attributesToKeep[name]){
-      target.removeAttribute(name);
-    }
-  }
-
-  updates.forEach(u => u());
-
-  return updates;
-};
-
-Template.prototype.loop = function($source, $target){
-
-  const $tempFragment = document.createDocumentFragment();
-
-  for(let i = 0, offset = 0;; i++){
-
-    const $sourceElement = $source.childNodes[i];
-    const $targetElement = $target.childNodes[i + offset];
-
-    //no further elements, end of loop
-    if(!$sourceElement && !$targetElement){
-      $target.appendChild($tempFragment);
-      break;
     }
 
-    //element doesn't exist anymore
-    if(!$sourceElement){
-      $target.removeChild($targetElement);
-      i--;
-      continue;
-    }
+    updates.forEach(u => u());
 
-    if(!$targetElement || !same($sourceElement, $targetElement, this.sanitizeNodePrefix)){
-      //@TODO strange behaviour here, have to make it a closure
-      const fn = ($target, $targetElement) => ($el) => {
-        if(!$targetElement){
-          //element should be newly created
-          return $tempFragment.appendChild($el);
-        }
-        //replace old node with new one
-        return $target.replaceChild($el, $targetElement);
-      };
-      const domFn = fn($target, $targetElement);
-      const getRange = ($targetElement, replacementType) => {
+    return updates;
+  };
 
-        if($targetElement && $targetElement.range && $targetElement.replacementType === replacementType && $targetElement.templateId === this.templateId){
-          return $targetElement.range;
-        } else {
-          const range = new NodesRange(document.createTextNode(''), document.createTextNode(''));
-          const { startNode } = range;
-          startNode.range = range;
-          startNode.replacementType = replacementType;
-          startNode.templateId = this.templateId;
-          domFn(range.extractContents());
-          return range;
-        }
+  Template.prototype.loop = function($source, $target){
+
+    const $tempFragment = document.createDocumentFragment();
+
+    const sourceChildren = $source.childNodes;
+
+    const getDomFn = ($targetElement) => {
+      if(!$targetElement){
+        //element should be newly created
+        return ($el) => $tempFragment.appendChild($el);
       }
-      switch($sourceElement.nodeType){
-        //regular DOM nodes
-        case NODE_TYPES.TEXT_NODE:
-          {
-            const content = $sourceElement.textContent;
-            const chunks = content.split(this.splitChunkRegex);
+      //replace old node with new one
+      return ($el) => $target.replaceChild($el, $targetElement);
+    };
 
-            const range = getRange($targetElement, 'textContent');
-            const processedChunks = this.processTextNodeChunks(chunks);
-            this.loop(processedChunks, range);
-            offset += range.childNodes.length + 1;
-            break;
+    for(let i = 0, offset = 0;; i++){
+
+      const $sourceElement = sourceChildren[i];
+      const $targetElement = $target.childNodes[i + offset];
+
+      //no further elements, end of loop
+      if(!$sourceElement && !$targetElement){
+        $target.appendChild($tempFragment);
+        break;
+      }
+
+      //element doesn't exist anymore
+      if(!$sourceElement){
+        $target.removeChild($targetElement);
+        i--;
+        continue;
+      }
+
+      if(!$targetElement || !same($sourceElement, $targetElement, sanitizeNodePrefix)){
+        const domFn = getDomFn($targetElement);
+        const getRange = ($targetElement, replacementType) => {
+
+          if($targetElement && $targetElement.range && $targetElement.replacementType === replacementType){
+            return $targetElement.range;
+          } else {
+            const range = new NodesRange(document.createTextNode(''), document.createTextNode(''));
+            const { startNode } = range;
+            startNode.range = range;
+            startNode.replacementType = replacementType;
+            domFn(range.extractContents());
+            return range;
           }
-        case NODE_TYPES.COMMENT_NODE:
-          domFn(document.createComment(this.replaceTokens($sourceElement.textContent)));
-          break;
-        case NODE_TYPES.ELEMENT_NODE:
-          const newChild = document.createElement($sourceElement.tagName.toLowerCase().replace(this.sanitizeNodePrefix, ''));
+        }
+        switch($sourceElement.nodeType){
+          //regular DOM nodes
+          case NODE_TYPES.TEXT_NODE:
+            {
+              const content = $sourceElement.textContent;
+              const chunks = content.split(splitChunkRegex);
 
-          this.loop($sourceElement, newChild);
-          domFn(newChild);
+              const range = getRange($targetElement, 'textContent');
+              const processedChunks = processTextNodeChunks(chunks);
+              this.loop(processedChunks, range);
+              offset += range.childNodes.length + 1;
+              break;
+            }
+          case NODE_TYPES.COMMENT_NODE:
+            domFn(document.createComment(replaceTokens($sourceElement.textContent, this.values)));
+            break;
+          case NODE_TYPES.ELEMENT_NODE:
+            const newChild = document.createElement($sourceElement.tagName.toLowerCase().replace(sanitizeNodePrefix, ''));
 
-          const updates = this.copyAttributes(newChild, $sourceElement);
-          this.updates = this.updates.concat(updates);
+            this.loop($sourceElement, newChild);
+            domFn(newChild);
 
-          break;
+            const updates = this.copyAttributes(newChild, $sourceElement);
+            this.updates = this.updates.concat(updates);
 
-        //custom nodes
-        case NODE_TYPES.TEXT_RESULT_NODE:
-          const chunk = `${$sourceElement.textContent || ''}`;
-          const match = chunk.match(this.matchChunkRegex);
-          if(match){
-            const range = getRange($targetElement, 'rangeInsertion'); //@TODO strange name
-            const updateFn = () => {
-              const oldValue = this.getChunkById(match[2], this.prevValues);
-              const newValue = this.getChunkById(match[2]);
+            break;
 
-              if(newValue === oldValue){
-                return;
+          //custom nodes
+          case NODE_TYPES.TEXT_RESULT_NODE:
+            const chunk = `${$sourceElement.textContent || ''}`;
+            const match = chunk.match(matchChunkRegex);
+            if(match){
+              const range = getRange($targetElement, 'rangeInsertion'); //@TODO strange name
+              const updateFn = () => {
+                const newValue = getChunkById(match[2], this.values);
+                const oldValue = getChunkById(match[2], this.prevValues);
+
+                if(newValue === oldValue){
+                  return;
+                }
+
+                const preocessedChunksContainer = copyTextNodeChunks(newValue);
+                this.loop(preocessedChunksContainer, range);
+              }
+              this.updates.push(updateFn);
+              updateFn();
+              offset += range.childNodes.length + 1;
+              break;
+            }
+            domFn(document.createTextNode($sourceElement.textContent));
+            break;
+          case NODE_TYPES.TEMPLATE_NODE:
+            {
+              const range = getRange($targetElement, 'template');
+
+              $sourceElement.template.render(range);
+              range.update();
+              offset += range.childNodes.length + 1;
+
+              break;
+            }
+          case NODE_TYPES.PROMISE_NODE:
+            {
+              const range = getRange($targetElement, 'futureResult');
+
+              $sourceElement.futureResult.then((response) => {
+                range.update();
+                const preocessedChunksContainer = copyTextNodeChunks(response);
+                this.loop(preocessedChunksContainer, range);
+                return preocessedChunksContainer;
+              });
+              offset += range.childNodes.length + 1;
+
+              break;
+            }
+          case NODE_TYPES.ELEMENT_RESULT_NODE:
+            {
+              const range = getRange($targetElement, 'element');
+              if($sourceElement.element instanceof DocumentFragment || range.childNodes.length !== 1){
+                range.childNodes.forEach(node => range.removeChild(node));
+                range.appendChild($sourceElement.element);
+              } else {
+                range.replaceChild($sourceElement.element, range.childNodes[0]);
               }
 
-              const preocessedChunksContainer = this.copyTextNodeChunks(newValue);
-              this.loop(preocessedChunksContainer, range);
+              offset += range.childNodes.length + 1;
+
+              break;
             }
-            this.updates.push(updateFn);
-            updateFn();
-            offset += range.childNodes.length + 1;
-            break;
-          }
-          domFn(document.createTextNode($sourceElement.textContent));
-          break;
-        case NODE_TYPES.TEMPLATE_NODE:
-          {
-            const range = getRange($targetElement, 'template');
+        }
 
-            $sourceElement.template.render(range);
-            range.update();
-            offset += range.childNodes.length + 1;
+        if($target instanceof NodesRange){
+          $target.update();
+        }
 
-            break;
-          }
-        case NODE_TYPES.PROMISE_NODE:
-          {
-            const range = getRange($targetElement, 'futureResult');
-
-            $sourceElement.futureResult.then((response) => {
-              range.update();
-              const preocessedChunksContainer = this.copyTextNodeChunks(response);
-              this.loop(preocessedChunksContainer, range);
-              return preocessedChunksContainer;
-            });
-            offset += range.childNodes.length + 1;
-
-            break;
-          }
-        case NODE_TYPES.ELEMENT_RESULT_NODE:
-          {
-            const range = getRange($targetElement, 'element');
-            if($sourceElement.element instanceof DocumentFragment || range.childNodes.length !== 1){
-              range.childNodes.forEach(node => range.removeChild(node));
-              range.appendChild($sourceElement.element);
-            } else {
-              range.replaceChild($sourceElement.element, range.childNodes[0]);
-            }
-
-            offset += range.childNodes.length + 1;
-
-            break;
-          }
+        continue;
       }
 
-      if($target instanceof NodesRange){
-        $target.update();
+      //at this point we are sure both elements exist
+
+      //same text
+      if(isSameTextNode($sourceElement, $targetElement)){
+        continue;
       }
 
-      continue;
+      //same node
+      if(same($sourceElement, $targetElement, sanitizeNodePrefix)){
+        this.loop($sourceElement, $targetElement);
+
+        const updates = this.copyAttributes($targetElement, $sourceElement);
+        this.updates = this.updates.concat(updates);
+        continue;
+      }
     }
+    return $target;
+  };
 
-    //at this point we are sure both elements exist
-
-    //same text
-    if(isSameTextNode($sourceElement, $targetElement)){
-      continue;
-    }
-
-    //same node
-    if(same($sourceElement, $targetElement, this.sanitizeNodePrefix)){
-      ($targetElement.nodeStopper !== stopNodeValue) && this.loop($sourceElement, $targetElement);
-
-      const updates = this.copyAttributes($targetElement, $sourceElement);
-      this.updates = this.updates.concat(updates);
-      continue;
-    }
+  return {
+    html,
+    render,
+    Template,
+    prepareLiterals,
+    generateContainer,
+    sanitize,
+    replaceTokens,
   }
-  return $target;
 };
 
-Template.prototype.render = function(target = document.createDocumentFragment()){
-  const cached = containersMap.get(target);
-  if((cached || {}).templateId === this.templateId){
-    cached.update(this.values);
-    return;
-  }
-  stopNode(target);
-  containersMap.set(target, {
-    templateId: this.templateId,
-    update: this.update.bind(this),
-  });
-  return this.loop(this.container, target);
-};
+const defaultInstance = createHtml({});
 
-Template.prototype.update = function(newData){
-  this.prevValues = this.values;
-  this.values = newData;
-  this.updates.forEach(u => u());
-};
+export const html = defaultInstance.html;
 
-Template.prototype.generateContainer = function(markup){
-  return this.parser.parseFromString(markup, "text/html").body;
-};
-
-Template.prototype.sanitize = function(str){
-  return str.replace(sanitizeTagsRegex, `<$1${this.sanitizeNodePrefix}$2`);
-};
-
-Template.prototype.prepareLiterals = function([firstChunk, ...restChunks]){
-  return restChunks.reduce((acc, chunk, index) => {
-    const keyName = this.generateTokenName(index);
-    return acc.concat(keyName).concat(chunk);
-  }, firstChunk);
-};
-
-Template.prototype.generateTokenName = function(index){
-  return `${this.PREFIX}${index}${this.POSTFIX}`;
-};
-
-Template.prototype.replaceTokens = function(text, dataMap = this.values){
-  return text.replace(this.replaceChunkRegex, (token, _, index) => {
-    const chunk = this.getChunkById(index, dataMap);
-    return isDefined(chunk) ? chunk : '';
-  });
-};
-
-Template.prototype.getTokenRegExp = function(groupMatches){
-  const indexRegex = `${groupMatches ? '(' : ''}\\d+${groupMatches ? ')' : ''}`;
-  return `(${regExpEscape(this.PREFIX)}${indexRegex}${regExpEscape(this.POSTFIX)})`;
-};
-
-
-export const html = (...args) => (new Template({})).parse(...args);
-
-export const render = (template, target) => template.render(target);
+export const render = defaultInstance.render;
 export const r = (...args) => render(html(...args));
+
+export const Template = defaultInstance.Template;
